Quick restructuring.
