+++ sim/src/bookinfo/details.rs
