+++ sim/src/bookinfo/mod.rs
@@ -1,0 +1,3 @@
+pub mod details;
+pub mod productpage;
+pub mod reviews;
