+++ sim/src/bookinfo/productpage.rs
