+++ sim/src/lib.rs
@@ -1,12 +1,10 @@
 #![feature(test)]
 #![feature(extern_types)]
-mod details;
+mod bookinfo;
 mod edge;
 mod filter_types;
 mod node;
 mod plugin_wrapper;
-mod productpage;
-mod reviews;
 mod sim_element;
 pub mod simulator;
 mod storage;
